requirements.txt
# Basic Django requirements
Django==4.2.10
python-dotenv==1.0.1

# Database
mysqlclient==2.2.3  # Se estiver usando MySQL
# psycopg2-binary==2.9.9  # Se estiver usando PostgreSQL

# PDF generation
reportlab==4.0.9
xhtml2pdf==0.2.13

# REST API
djangorestframework==3.14.0

# Authentication and security
django-allauth==0.61.1
django-otp==1.3.0

# Internationalization
django-rosetta==0.10.0

# Development tools
django-debug-toolbar==4.3.0

# Production
gunicorn==21.2.0
whitenoise==6.6.0

# History and auditing
django-simple-history==3.4.0